+++ what_team_my_team/app/(profile)/profile/[slug]/(profile)/_components/TechInfo.tsx
